Prdt 136 2 (#169)
* PRDT-136: imporve local CDK build times


* PRDT-136: fix tests for local build changes


---------
