+++ lib/handlers/activities/_acCollectionId/GET/index.js
@@ -2,9 +2,9 @@ const {
   getActivitiesByAcCollectionId,
   getActivitiesByActivityType,
   getActivityByActivityId,
-} = require("/opt/activities/methods");
+} = require("../../methods");
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { ALLOWED_FILTERS } = require("/opt/activities/configs");
+const { ALLOWED_FILTERS } = require("../../configs");
 
 /**
  * @api {get} /activities/{acCollectionId} GET
