+++ lib/handlers/activities/_acCollectionId/POST/index.js
@@ -1,7 +1,7 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
 const { quickApiPutHandler } = require("/opt/data-utils");
-const { ACTIVITY_API_PUT_CONFIG } = require("/opt/activities/configs");
-const { parseRequest } = require("/opt/activities/methods");
+const { ACTIVITY_API_PUT_CONFIG } = require("../../configs");
+const { parseRequest } = require("../../methods");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
 
 /**
