+++ lib/handlers/activities/_acCollectionId/PUT/index.js
@@ -1,7 +1,7 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
 const { quickApiUpdateHandler } = require("/opt/data-utils");
-const { ACTIVITY_API_UPDATE_CONFIG } = require("/opt/activities/configs");
-const { parseRequest } = require("/opt/activities/methods");
+const { ACTIVITY_API_UPDATE_CONFIG } = require("../../configs");
+const { parseRequest } = require("../../methods");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
 
 /**
