+++ lib/handlers/activities/configs.js
