+++ lib/handlers/activities/methods.js
@@ -1,7 +1,6 @@
-const { TABLE_NAME, batchGetData, runQuery, getOne, marshall, incrementCounter } = require("/opt/dynamodb");
+const { TABLE_NAME, runQuery, getOne, marshall, incrementCounter } = require("/opt/dynamodb");
 const { Exception, logger } = require("/opt/base");
-const { ALLOWED_FILTERS } = require("/opt/facilities/configs");
-
+const { ALLOWED_FILTERS } = require("./configs");
 /**
  * Adds any filter expressions if any filters were added to query.
  *
@@ -20,7 +19,6 @@ function addFilters(queryObj, filters) {
         } else if (!queryObj.FilterExpression) {
           queryObj.FilterExpression = "";
         }
-
         if (!queryObj.ExpressionAttributeNames) {
           queryObj.ExpressionAttributeNames = {};
         }
@@ -45,25 +43,25 @@ function addFilters(queryObj, filters) {
 }
 
 /**
- * Retrieves all facilities matching a facility collection id.
+ * Retrieves all activities matching an Activity Collection ID.
  *
  * @async
- * @param {Object} fcCollectionId - The fcCollectionId of the facility.
+ * @param {Object} acCollectionId - The Activity Collection ID of the activity.
  * @param {Object} [params] - Optional parameters for pagination control.
  * @param {number} [params.limit] - Maximum number of items to return.
  * @param {string} [params.lastEvaluatedKey] - Key to resume pagination from.
  * @param {boolean} [params.paginated=true] - Whether to enable pagination.
  *
  * @returns {Promise<Object>} Query response containing:
- *   - items: Array of facility objects
+ *   - items: Array of activity objects
  *   - lastEvaluatedKey: Key for pagination continuation
  *   - count: Number of items returned
  *
  * @throws {Exception} With code 400 if database operation fails
  *
  */
-async function getFacilitiesByFcCollectionId(fcCollectionId, filters, params = null) {
-  logger.info("Get Facilities by Facility Collection ID");
+async function getActivitiesByAcCollectionId(acCollectionId, filters, params = null) {
+  logger.info("Get Activities by Activity Collection ID");
   try {
     const limit = params?.limit || null;
     const lastEvaluatedKey = params?.lastEvaluatedKey || null;
@@ -72,7 +70,7 @@ async function getFacilitiesByFcCollectionId(fcCollectionId, filters, params = n
       TableName: TABLE_NAME,
       KeyConditionExpression: "pk = :pk",
       ExpressionAttributeValues: {
-        ":pk": { S: `facility::${fcCollectionId}` },
+        ":pk": { S: `activity::${acCollectionId}` },
       },
     };
 
@@ -81,10 +79,10 @@ async function getFacilitiesByFcCollectionId(fcCollectionId, filters, params = n
     }
 
     const res = await runQuery(queryObj, limit, lastEvaluatedKey, paginated);
-    logger.info(`Facilities: ${res?.items?.length} found.`);
+    logger.info(`Activities: ${res?.items?.length} found.`);
     return res;
   } catch (error) {
-    throw new Exception("Error getting facilities", {
+    throw new Exception("Error getting activities", {
       code: 400,
       error: error,
     });
@@ -92,30 +90,30 @@ async function getFacilitiesByFcCollectionId(fcCollectionId, filters, params = n
 }
 
 /**
- * Retrieves facilities matching the specified Facility Type.
+ * Retrieves activities matching the specified activity type.
  *
  * @async
- * @param {Object} fcCollectionId - The facility collection ID of the facility.
- * @param {Object} facilityType - Type of facility to retrieve.
+ * @param {Object} acCollectionId - The Activity Collection Id of the activity.
+ * @param {Object} activityType - Type of activity to retrieve.
  * @param {Object} [params] - Optional parameters for pagination control.
  * @param {number} [params.limit] - Maximum number of items to return.
  * @param {string} [params.lastEvaluatedKey] - Key to resume pagination from.
  * @param {boolean} [params.paginated=true] - Whether to enable pagination.
  *
  * @returns {Promise<Object>} Query response containing:
- *   - items: Array of facility objects
+ *   - items: Array of activity objects
  *   - lastEvaluatedKey: Key for pagination continuation
  *   - count: Number of items returned
  *
  * @throws {Exception} With code 400 if database operation fails
  */
-async function getFacilitiesByFacilityType(
-  fcCollectionId,
-  facilityType,
+async function getActivitiesByActivityType(
+  acCollectionId,
+  activityType,
   filters,
   params = null
 ) {
-  logger.info("Get Facility by Facility Type");
+  logger.info("Get Activity by activity type");
   try {
     const limit = params?.limit || null;
     const lastEvaluatedKey = params?.lastEvaluatedKey || null;
@@ -124,8 +122,8 @@ async function getFacilitiesByFacilityType(
       TableName: TABLE_NAME,
       KeyConditionExpression: "pk = :pk AND begins_with(sk, :sk)",
       ExpressionAttributeValues: {
-        ":pk": { S: `facility::${fcCollectionId}` },
-        ":sk": { S: `${facilityType}::` },
+        ":pk": { S: `activity::${acCollectionId}` },
+        ":sk": { S: `${activityType}::` },
       },
     };
 
@@ -134,10 +132,10 @@ async function getFacilitiesByFacilityType(
     }
 
     const res = await runQuery(queryObj, limit, lastEvaluatedKey, paginated);
-    logger.info(`Facilities: ${res?.items?.length} found.`);
+    logger.info(`Activities: ${res?.items?.length} found.`);
     return res;
   } catch (error) {
-    throw new Exception("Error getting facilities", {
+    throw new Exception("Error getting activities", {
       code: 400,
       error: error,
     });
@@ -145,61 +143,60 @@ async function getFacilitiesByFacilityType(
 }
 
 /**
- * Retrieves facilities matching the specified Facility ID.
+ * Retrieves activities matching the specified activity type.
  *
  * @async
- * @param {Object} fcCollectionId - The facility collection ID number.
- * @param {Object} facilityType - Type of facility.
- * @param {Object} facilityId - The ID of the facility.
+ * @param {Object} acCollectionId - The Activity Collection ID of the activity.
+ * @param {Object} activityType - Type of activity to retrieve.
+ * @param {Object} activityId - The ID of the activity to retrieve.
  *
  * @returns {Promise<Object>} Query response containing:
- *   - items: facility object
+ *   - items: activity object
  *
  * @throws {Exception} With code 400 if database operation fails
  */
-async function getFacilityByFacilityId(fcCollectionId, facilityType, facilityId, fetchActivities = false) {
-  logger.info("Get Facility By Facility Type and ID");
+async function getActivityByActivityId(acCollectionId, activityType, activityId, fetchGeozone = false) {
+  logger.info("Get Activity By activity type and ID");
   try {
     let res = await getOne(
-      `facility::${fcCollectionId}`,
-      `${facilityType}::${facilityId}`
+      `activity::${acCollectionId}`,
+      `${activityType}::${activityId}`
     );
-    if (fetchActivities && res?.activities?.length) {
-      logger.debug(`Fetching activities: ${JSON.stringify(res.activities, null, 2)}`);
-      // Batch get the activities and staple them to the response
-      res.activities = await batchGetData(res.activities, TABLE_NAME);
+    if (fetchGeozone && res?.geozone?.pk && res?.geozone?.sk) {
+      const geozone = await getOne(res.geozone.pk, res.geozone.sk);
+      res.geozone = geozone;
     }
-    logger.debug(`Facility: ${JSON.stringify(res, null, 2)}`);
     return res;
+
   } catch (error) {
-    throw new Exception("Error getting facility", { code: 400, error: error });
+    throw new Exception("Error getting activity", { code: 400, error: error });
   }
 }
 
 /**
  * Processes incoming requests by handling batch operations and individual items.
- * Validates facilityType and facilityId parameters, falling back to body values if needed.
+ * Validates activityType and activityId parameters, falling back to body values if needed.
  *
- * @param {Object} fcCollectionId - Facility Collection ID number
+ * @param {Object} acCollectionId - Activity Collection ID number
  * @param {Object|Array} body - Request payload (single item or array of items)
  * @param {string} requestType - Type of HTTP request ("PUT" or "POST")
- * @param {string} facilityType - Type of facility
- * @param {string} [facilityId] - Optional ID of the facility
+ * @param {string} activityType - Type of activity
+ * @param {string} [activityId] - Optional ID of the activity
  *
  * @returns {Array} Array of processed update requests
  */
-async function parseRequest(fcCollectionId, body, requestType, facilityType, facilityId = null) {
+async function parseRequest(acCollectionId, body, requestType, activityType, activityId = null) {
   let updateRequests = [];
   // Check if the request is a batch request
   if (Array.isArray(body)) {
     for (let item of body) {
       updateRequests.push(
-        await processItem(fcCollectionId, item, requestType, facilityType, facilityId)
+        await processItem(acCollectionId, item, requestType, activityType, activityId)
       );
     }
   } else {
     updateRequests.push(
-      await processItem(fcCollectionId, body, requestType, facilityType, facilityId)
+      await processItem(acCollectionId, body, requestType, activityType, activityId)
     );
   }
 
@@ -210,68 +207,58 @@ async function parseRequest(fcCollectionId, body, requestType, facilityType, fac
  * Processes individual items based on request type, creating appropriate key structures.
  * Handles PUT and POST requests differently, managing primary and sort keys accordingly.
  *
- * @param {Object} fcCollectionId - Facility Collection ID number
+ * @param {Object} acCollectionId - Activity Collection ID number
  * @param {Object} item - Item data to process
  * @param {string} requestType - Type of HTTP request ("PUT" or "POST")
- * @param {string} facilityType - Type of facility
- * @param {string} [facilityId] - Optional ID of the facility
+ * @param {string} activityType - Type of activity
+ * @param {string} [activityId] - Optional ID of the activity
  *
  * @returns {Object} Processed item with key structure and data
  */
 async function processItem(
-  fcCollectionId,
+  acCollectionId,
   item,
   requestType,
-  facilityType,
-  facilityId = null
+  activityType,
+  activityId = null
 ) {
-  const pk = `facility::${fcCollectionId}`;
-  let sk = null;
+  const pk = `activity::${acCollectionId}`;
+  let sk = null
 
-  // You should have a facilityType in queryParams or the item
-  if (!facilityType && !item.facilityType) {
-    throw new Error(`facilityType is not specified in one of the requests.`, { code: 400 });
+  // You should have a activityType in queryParams or the item
+  if (!activityType && !item.activityType) {
+    throw new Error(`activityType is not specified in one of the requests.`, { code: 400 });
   }
 
-  // facilityType should be taken from queryParams, but can be taken from item if it's a bulk update
-  facilityType = facilityType ?? item?.facilityType;
+  // activityType should be taken from queryParams, but can be taken from item if it's a bulk update
+  activityType = activityType ?? item?.activityType;
 
   if (requestType == "PUT") {
-    // Throw an error if facilityId is not passed in
-    if (!facilityId && !item.facilityId) {
-      throw new Error(`facilityId is not specified in one of the requests.`, { code: 400 });
+    // Throw an error if activityId is not passed in
+    if (!activityId && !item.activityId) {
+      throw new Error(`activityId is not specified in one of the requests.`, { code: 400 });
     }
 
-    // facilityId should be taken from queryParams, but can be taken from item if it's a bulk update
-    facilityId = facilityId ?? item?.facilityId;
-    sk = `${facilityType}::${facilityId}`;
+    // activityId should be taken from queryParams, but can be taken from item if it's a bulk update
+    activityId = activityId ?? item?.activityId;
+    sk = `${activityType}::${activityId}`;
 
     // Remove items that can't be in a PUT request
     delete item.pk;
     delete item.sk;
-    delete item.facilityType;
-    delete item.facilityId;
+    delete item.activityType;
+    delete item.activityId;
   } else if (requestType == "POST") {
-    // Throw an error if facilityId is passed in POST request, as we only allow auto increment for POST requests
-    if (facilityId) {
+    // Throw an error if activityId is passed in POST request, as we only allow auto increment for POST requests
+    if (activityId) {
       throw new Error(
-        "Can't specify facilityId in POST request; must be null to allow auto increment",
+        "Can't specify activityId in POST request; must be null to allow auto increment",
         { code: 400 }
       );
     }
 
-    // Make sure defaults are set for the item
-    item['fcCollectionId'] = fcCollectionId;
-    item['facilityType'] = facilityType;
-    if (!item?.hasOwnProperty('activities')) {
-      item.activities = [];
-    }
-    if (!item?.hasOwnProperty('showOnMap')) {
-      item.showOnMap = true;
-    }
-
     // If it's a retry attempt, remove these attributes from the item (if they exist)
-    delete item?.facilityId;
+    delete item?.activityId;
     delete item?.creationDate;
     delete item?.lastUpdated;
     delete item?.version;
@@ -279,15 +266,15 @@ async function processItem(
     // If it's a POST request, we need to increment the identifier
     // or we need to create the counter if it doesn't already exist. incrementCounter does either.
     // Pass in the pk to start iterating the collection.
-    const identifier = await incrementCounter(pk, [facilityType]);
+    const identifier = await incrementCounter(pk, [activityType]);
 
-    // Create the sk from the facilityType and the identifier
-    sk = `${facilityType}::${String(identifier)}`;
+    // Create the sk from the activityType and the identifier
+    sk = `${activityType}::${String(identifier)}`;
 
     item.pk = pk;
     item.sk = sk;
-    item.facilityType = facilityType;
-    item.facilityId = Number(identifier);
+    item.activityType = activityType;
+    item.activityId = Number(identifier);
     item.identifier = Number(identifier);
   }
 
@@ -298,9 +285,9 @@ async function processItem(
 }
 
 module.exports = {
-  getFacilitiesByFcCollectionId,
-  getFacilitiesByFacilityType,
-  getFacilityByFacilityId,
+  getActivitiesByAcCollectionId,
+  getActivitiesByActivityType,
+  getActivityByActivityId,
   parseRequest,
   processItem
 };
