+++ lib/handlers/activities/resources.js
@@ -18,8 +18,8 @@ function activitiesSetup(scope, props) {
 
   // GET /activities/{acCollectionId}
   const activitiesGetByAcCollectionId = new NodejsFunction(scope, 'ActivitiesGetByAcCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/activities/_acCollectionId/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/activities'),
+    handler: '_acCollectionId/GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -28,8 +28,8 @@ function activitiesSetup(scope, props) {
 
   // POST /activities/{acCollectionId}
   const activitiesPostByAcCollectionId = new NodejsFunction(scope, 'ActivitiesPostByAcCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/activities/_acCollectionId/POST'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/activities'),
+    handler: '_acCollectionId/POST/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -38,8 +38,8 @@ function activitiesSetup(scope, props) {
 
   // PUT /activities/{acCollectionId}/{activityType}/{activityId}
   const activitiesPutByCollectionTypeId = new NodejsFunction(scope, 'ActivitiesPutByCollectionTypeId', {
-    code: lambda.Code.fromAsset('lib/handlers/activities/_acCollectionId/PUT'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/activities'),
+    handler: '_acCollectionId/PUT/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -48,8 +48,8 @@ function activitiesSetup(scope, props) {
 
   // DELETE /activities/{acCollectionId}/{activityType}/{activityId}
   const activitiesDeleteByCollectionTypeId = new NodejsFunction(scope, 'ActivitiesDeleteByCollectionTypeId', {
-    code: lambda.Code.fromAsset('lib/handlers/activities/_acCollectionId/DELETE'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/activities'),
+    handler: '_acCollectionId/DELETE/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
