+++ lib/handlers/bookings/GET/index.js
@@ -1,7 +1,7 @@
 // Get booking
 
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { getBookingByBookingId, getBookingsByActivityDetails, getBookingsByUserSub } = require("/opt/bookings/methods");
+const { getBookingByBookingId, getBookingsByActivityDetails, getBookingsByUserSub } = require("../methods");
 
 exports.handler = async (event, context) => {
   logger.info("Bookings GET:", event);
