+++ lib/handlers/bookings/POST/index.js
@@ -1,7 +1,7 @@
 // Create new booking
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { createBooking } = require("/opt/bookings/methods");
-const { BOOKING_PUT_CONFIG } = require("/opt/bookings/configs");
+const { createBooking } = require("../methods");
+const { BOOKING_PUT_CONFIG } = require("../configs");
 const { quickApiPutHandler } = require("/opt/data-utils");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
