+++ lib/handlers/bookings/PUT/index.js
@@ -1,7 +1,7 @@
 // Create new transaction
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { confirmBooking } = require("/opt/bookings/methods")
-const { BOOKING_UPDATE_CONFIG } = require("/opt/bookings/configs");
+const { confirmBooking } = require("../methods")
+const { BOOKING_UPDATE_CONFIG } = require("../configs");
 const { quickApiUpdateHandler } = require("/opt/data-utils");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
