+++ lib/handlers/bookings/methods.js
@@ -7,8 +7,8 @@ const {
   TABLE_NAME,
 } = require("/opt/dynamodb");
 const { Exception, logger } = require("/opt/base");
-const { getActivityByActivityId } = require("/opt/activities/methods");
-const { getAndAttachNestedProperties } = require("../data-utils");
+const { getActivityByActivityId } = require("../activities/methods");
+const { getAndAttachNestedProperties } = require("/opt/data-utils");
 
 async function getBookingByBookingId(bookingId, fetchAccessPoints = false) {
   logger.debug("Getting booking by bookingId:", bookingId);
