+++ lib/handlers/bookings/resources.js
@@ -31,8 +31,8 @@ function bookingsSetup(scope, props) {
 
   // GET /bookings
   const bookingsGet = new NodejsFunction(scope, 'BookingsGet', {
-    code: lambda.Code.fromAsset('lib/handlers/bookings/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers'),
+    handler: 'bookings/GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -43,8 +43,8 @@ function bookingsSetup(scope, props) {
 
   // POST /bookings
   const bookingsPost = new NodejsFunction(scope, 'BookingsPost', {
-    code: lambda.Code.fromAsset('lib/handlers/bookings/POST'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers'),
+    handler: 'bookings/POST/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -55,8 +55,8 @@ function bookingsSetup(scope, props) {
  
   // PUT /bookings
   const bookingsPut = new NodejsFunction(scope, 'BookingsPut', {
-    code: lambda.Code.fromAsset('lib/handlers/bookings/PUT'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers'),
+    handler: 'bookings/PUT/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
