+++ lib/handlers/facilities/_fcCollectionId/GET/index.js
@@ -2,9 +2,9 @@ const {
   getFacilitiesByFcCollectionId,
   getFacilitiesByFacilityType,
   getFacilityByFacilityId,
-} = require("/opt/facilities/methods");
+} = require("../../methods");
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { ALLOWED_FILTERS } = require("/opt/facilities/configs");
+const { ALLOWED_FILTERS } = require("../../configs");
 
 /**
  * @api {get} /facilities/{fcCollectionId} GET
