+++ lib/handlers/facilities/_fcCollectionId/PUT/index.js
@@ -1,7 +1,7 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
 const { quickApiUpdateHandler } = require("/opt/data-utils");
-const { FACILITY_API_UPDATE_CONFIG } = require("/opt/facilities/configs");
-const { parseRequest } = require("/opt/facilities/methods");
+const { FACILITY_API_UPDATE_CONFIG } = require("../../configs");
+const { parseRequest } = require("../../methods");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
 
 /**
