+++ lib/handlers/facilities/methods.js
@@ -1,6 +1,6 @@
-const { TABLE_NAME, runQuery, getOne, marshall, incrementCounter } = require("/opt/dynamodb");
+const { TABLE_NAME, batchGetData, runQuery, getOne, marshall, incrementCounter } = require("/opt/dynamodb");
 const { Exception, logger } = require("/opt/base");
-const { ALLOWED_FILTERS } = require("/opt/activities/configs");
+const { ALLOWED_FILTERS } = require("./configs");
 
 /**
  * Adds any filter expressions if any filters were added to query.
@@ -20,6 +20,7 @@ function addFilters(queryObj, filters) {
         } else if (!queryObj.FilterExpression) {
           queryObj.FilterExpression = "";
         }
+
         if (!queryObj.ExpressionAttributeNames) {
           queryObj.ExpressionAttributeNames = {};
         }
@@ -44,25 +45,25 @@ function addFilters(queryObj, filters) {
 }
 
 /**
- * Retrieves all activities matching an Activity Collection ID.
+ * Retrieves all facilities matching a facility collection id.
  *
  * @async
- * @param {Object} acCollectionId - The Activity Collection ID of the activity.
+ * @param {Object} fcCollectionId - The fcCollectionId of the facility.
  * @param {Object} [params] - Optional parameters for pagination control.
  * @param {number} [params.limit] - Maximum number of items to return.
  * @param {string} [params.lastEvaluatedKey] - Key to resume pagination from.
  * @param {boolean} [params.paginated=true] - Whether to enable pagination.
  *
  * @returns {Promise<Object>} Query response containing:
- *   - items: Array of activity objects
+ *   - items: Array of facility objects
  *   - lastEvaluatedKey: Key for pagination continuation
  *   - count: Number of items returned
  *
  * @throws {Exception} With code 400 if database operation fails
  *
  */
-async function getActivitiesByAcCollectionId(acCollectionId, filters, params = null) {
-  logger.info("Get Activities by Activity Collection ID");
+async function getFacilitiesByFcCollectionId(fcCollectionId, filters, params = null) {
+  logger.info("Get Facilities by Facility Collection ID");
   try {
     const limit = params?.limit || null;
     const lastEvaluatedKey = params?.lastEvaluatedKey || null;
@@ -71,7 +72,7 @@ async function getActivitiesByAcCollectionId(acCollectionId, filters, params = n
       TableName: TABLE_NAME,
       KeyConditionExpression: "pk = :pk",
       ExpressionAttributeValues: {
-        ":pk": { S: `activity::${acCollectionId}` },
+        ":pk": { S: `facility::${fcCollectionId}` },
       },
     };
 
@@ -80,10 +81,10 @@ async function getActivitiesByAcCollectionId(acCollectionId, filters, params = n
     }
 
     const res = await runQuery(queryObj, limit, lastEvaluatedKey, paginated);
-    logger.info(`Activities: ${res?.items?.length} found.`);
+    logger.info(`Facilities: ${res?.items?.length} found.`);
     return res;
   } catch (error) {
-    throw new Exception("Error getting activities", {
+    throw new Exception("Error getting facilities", {
       code: 400,
       error: error,
     });
@@ -91,30 +92,30 @@ async function getActivitiesByAcCollectionId(acCollectionId, filters, params = n
 }
 
 /**
- * Retrieves activities matching the specified activity type.
+ * Retrieves facilities matching the specified Facility Type.
  *
  * @async
- * @param {Object} acCollectionId - The Activity Collection Id of the activity.
- * @param {Object} activityType - Type of activity to retrieve.
+ * @param {Object} fcCollectionId - The facility collection ID of the facility.
+ * @param {Object} facilityType - Type of facility to retrieve.
  * @param {Object} [params] - Optional parameters for pagination control.
  * @param {number} [params.limit] - Maximum number of items to return.
  * @param {string} [params.lastEvaluatedKey] - Key to resume pagination from.
  * @param {boolean} [params.paginated=true] - Whether to enable pagination.
  *
  * @returns {Promise<Object>} Query response containing:
- *   - items: Array of activity objects
+ *   - items: Array of facility objects
  *   - lastEvaluatedKey: Key for pagination continuation
  *   - count: Number of items returned
  *
  * @throws {Exception} With code 400 if database operation fails
  */
-async function getActivitiesByActivityType(
-  acCollectionId,
-  activityType,
+async function getFacilitiesByFacilityType(
+  fcCollectionId,
+  facilityType,
   filters,
   params = null
 ) {
-  logger.info("Get Activity by activity type");
+  logger.info("Get Facility by Facility Type");
   try {
     const limit = params?.limit || null;
     const lastEvaluatedKey = params?.lastEvaluatedKey || null;
@@ -123,8 +124,8 @@ async function getActivitiesByActivityType(
       TableName: TABLE_NAME,
       KeyConditionExpression: "pk = :pk AND begins_with(sk, :sk)",
       ExpressionAttributeValues: {
-        ":pk": { S: `activity::${acCollectionId}` },
-        ":sk": { S: `${activityType}::` },
+        ":pk": { S: `facility::${fcCollectionId}` },
+        ":sk": { S: `${facilityType}::` },
       },
     };
 
@@ -133,10 +134,10 @@ async function getActivitiesByActivityType(
     }
 
     const res = await runQuery(queryObj, limit, lastEvaluatedKey, paginated);
-    logger.info(`Activities: ${res?.items?.length} found.`);
+    logger.info(`Facilities: ${res?.items?.length} found.`);
     return res;
   } catch (error) {
-    throw new Exception("Error getting activities", {
+    throw new Exception("Error getting facilities", {
       code: 400,
       error: error,
     });
@@ -144,60 +145,61 @@ async function getActivitiesByActivityType(
 }
 
 /**
- * Retrieves activities matching the specified activity type.
+ * Retrieves facilities matching the specified Facility ID.
  *
  * @async
- * @param {Object} acCollectionId - The Activity Collection ID of the activity.
- * @param {Object} activityType - Type of activity to retrieve.
- * @param {Object} activityId - The ID of the activity to retrieve.
+ * @param {Object} fcCollectionId - The facility collection ID number.
+ * @param {Object} facilityType - Type of facility.
+ * @param {Object} facilityId - The ID of the facility.
  *
  * @returns {Promise<Object>} Query response containing:
- *   - items: activity object
+ *   - items: facility object
  *
  * @throws {Exception} With code 400 if database operation fails
  */
-async function getActivityByActivityId(acCollectionId, activityType, activityId, fetchGeozone = false) {
-  logger.info("Get Activity By activity type and ID");
+async function getFacilityByFacilityId(fcCollectionId, facilityType, facilityId, fetchActivities = false) {
+  logger.info("Get Facility By Facility Type and ID");
   try {
     let res = await getOne(
-      `activity::${acCollectionId}`,
-      `${activityType}::${activityId}`
+      `facility::${fcCollectionId}`,
+      `${facilityType}::${facilityId}`
     );
-    if (fetchGeozone && res?.geozone?.pk && res?.geozone?.sk) {
-      const geozone = await getOne(res.geozone.pk, res.geozone.sk);
-      res.geozone = geozone;
+    if (fetchActivities && res?.activities?.length) {
+      logger.debug(`Fetching activities: ${JSON.stringify(res.activities, null, 2)}`);
+      // Batch get the activities and staple them to the response
+      res.activities = await batchGetData(res.activities, TABLE_NAME);
     }
+    logger.debug(`Facility: ${JSON.stringify(res, null, 2)}`);
     return res;
-
   } catch (error) {
-    throw new Exception("Error getting activity", { code: 400, error: error });
+    throw new Exception("Error getting facility", { code: 400, error: error });
   }
 }
 
 /**
  * Processes incoming requests by handling batch operations and individual items.
- * Validates activityType and activityId parameters, falling back to body values if needed.
+ * Validates facilityType and facilityId parameters, falling back to body values if needed.
  *
- * @param {Object} acCollectionId - Activity Collection ID number
+ * @param {Object} fcCollectionId - Facility Collection ID number
  * @param {Object|Array} body - Request payload (single item or array of items)
  * @param {string} requestType - Type of HTTP request ("PUT" or "POST")
- * @param {string} activityType - Type of activity
- * @param {string} [activityId] - Optional ID of the activity
+ * @param {string} facilityType - Type of facility
+ * @param {string} [facilityId] - Optional ID of the facility
  *
  * @returns {Array} Array of processed update requests
  */
-async function parseRequest(acCollectionId, body, requestType, activityType, activityId = null) {
+async function parseRequest(fcCollectionId, body, requestType, facilityType, facilityId = null) {
   let updateRequests = [];
   // Check if the request is a batch request
   if (Array.isArray(body)) {
     for (let item of body) {
       updateRequests.push(
-        await processItem(acCollectionId, item, requestType, activityType, activityId)
+        await processItem(fcCollectionId, item, requestType, facilityType, facilityId)
       );
     }
   } else {
     updateRequests.push(
-      await processItem(acCollectionId, body, requestType, activityType, activityId)
+      await processItem(fcCollectionId, body, requestType, facilityType, facilityId)
     );
   }
 
@@ -208,58 +210,68 @@ async function parseRequest(acCollectionId, body, requestType, activityType, act
  * Processes individual items based on request type, creating appropriate key structures.
  * Handles PUT and POST requests differently, managing primary and sort keys accordingly.
  *
- * @param {Object} acCollectionId - Activity Collection ID number
+ * @param {Object} fcCollectionId - Facility Collection ID number
  * @param {Object} item - Item data to process
  * @param {string} requestType - Type of HTTP request ("PUT" or "POST")
- * @param {string} activityType - Type of activity
- * @param {string} [activityId] - Optional ID of the activity
+ * @param {string} facilityType - Type of facility
+ * @param {string} [facilityId] - Optional ID of the facility
  *
  * @returns {Object} Processed item with key structure and data
  */
 async function processItem(
-  acCollectionId,
+  fcCollectionId,
   item,
   requestType,
-  activityType,
-  activityId = null
+  facilityType,
+  facilityId = null
 ) {
-  const pk = `activity::${acCollectionId}`;
-  let sk = null
+  const pk = `facility::${fcCollectionId}`;
+  let sk = null;
 
-  // You should have a activityType in queryParams or the item
-  if (!activityType && !item.activityType) {
-    throw new Error(`activityType is not specified in one of the requests.`, { code: 400 });
+  // You should have a facilityType in queryParams or the item
+  if (!facilityType && !item.facilityType) {
+    throw new Error(`facilityType is not specified in one of the requests.`, { code: 400 });
   }
 
-  // activityType should be taken from queryParams, but can be taken from item if it's a bulk update
-  activityType = activityType ?? item?.activityType;
+  // facilityType should be taken from queryParams, but can be taken from item if it's a bulk update
+  facilityType = facilityType ?? item?.facilityType;
 
   if (requestType == "PUT") {
-    // Throw an error if activityId is not passed in
-    if (!activityId && !item.activityId) {
-      throw new Error(`activityId is not specified in one of the requests.`, { code: 400 });
+    // Throw an error if facilityId is not passed in
+    if (!facilityId && !item.facilityId) {
+      throw new Error(`facilityId is not specified in one of the requests.`, { code: 400 });
     }
 
-    // activityId should be taken from queryParams, but can be taken from item if it's a bulk update
-    activityId = activityId ?? item?.activityId;
-    sk = `${activityType}::${activityId}`;
+    // facilityId should be taken from queryParams, but can be taken from item if it's a bulk update
+    facilityId = facilityId ?? item?.facilityId;
+    sk = `${facilityType}::${facilityId}`;
 
     // Remove items that can't be in a PUT request
     delete item.pk;
     delete item.sk;
-    delete item.activityType;
-    delete item.activityId;
+    delete item.facilityType;
+    delete item.facilityId;
   } else if (requestType == "POST") {
-    // Throw an error if activityId is passed in POST request, as we only allow auto increment for POST requests
-    if (activityId) {
+    // Throw an error if facilityId is passed in POST request, as we only allow auto increment for POST requests
+    if (facilityId) {
       throw new Error(
-        "Can't specify activityId in POST request; must be null to allow auto increment",
+        "Can't specify facilityId in POST request; must be null to allow auto increment",
         { code: 400 }
       );
     }
 
+    // Make sure defaults are set for the item
+    item['fcCollectionId'] = fcCollectionId;
+    item['facilityType'] = facilityType;
+    if (!item?.hasOwnProperty('activities')) {
+      item.activities = [];
+    }
+    if (!item?.hasOwnProperty('showOnMap')) {
+      item.showOnMap = true;
+    }
+
     // If it's a retry attempt, remove these attributes from the item (if they exist)
-    delete item?.activityId;
+    delete item?.facilityId;
     delete item?.creationDate;
     delete item?.lastUpdated;
     delete item?.version;
@@ -267,15 +279,15 @@ async function processItem(
     // If it's a POST request, we need to increment the identifier
     // or we need to create the counter if it doesn't already exist. incrementCounter does either.
     // Pass in the pk to start iterating the collection.
-    const identifier = await incrementCounter(pk, [activityType]);
+    const identifier = await incrementCounter(pk, [facilityType]);
 
-    // Create the sk from the activityType and the identifier
-    sk = `${activityType}::${String(identifier)}`;
+    // Create the sk from the facilityType and the identifier
+    sk = `${facilityType}::${String(identifier)}`;
 
     item.pk = pk;
     item.sk = sk;
-    item.activityType = activityType;
-    item.activityId = Number(identifier);
+    item.facilityType = facilityType;
+    item.facilityId = Number(identifier);
     item.identifier = Number(identifier);
   }
 
@@ -286,9 +298,9 @@ async function processItem(
 }
 
 module.exports = {
-  getActivitiesByAcCollectionId,
-  getActivitiesByActivityType,
-  getActivityByActivityId,
+  getFacilitiesByFcCollectionId,
+  getFacilitiesByFacilityType,
+  getFacilityByFacilityId,
   parseRequest,
   processItem
 };
