+++ lib/handlers/facilities/resources.js
@@ -21,8 +21,8 @@ function facilitiesSetup(scope, props) {
 
   // GET /facilities/{fcCollectionId}
   const facilitiesGetByFcCollectionId = new NodejsFunction(scope, 'FacilitiesGetByFcCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/facilities/_fcCollectionId/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/facilities'),
+    handler: '_fcCollectionId/GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -33,8 +33,8 @@ function facilitiesSetup(scope, props) {
 
   // POST /facilities/{fcCollectionId}
   const facilitiesPostByFcCollectionId = new NodejsFunction(scope, 'FacilitiesPostByFcCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/facilities/_fcCollectionId/POST'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/facilities'),
+    handler: '_fcCollectionId/POST/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -44,8 +44,8 @@ function facilitiesSetup(scope, props) {
 
   // PUT /facilities/{fcCollectionId}
   const facilitiesPutByFcCollectionId = new NodejsFunction(scope, 'FacilitiesPutByFcCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/facilities/_fcCollectionId/PUT'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/facilities'),
+    handler: '_fcCollectionId/PUT/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -56,8 +56,8 @@ function facilitiesSetup(scope, props) {
 
   // DELETE /facilities/{fcCollectionId}
   const facilitiesDeleteByFcCollectionId = new NodejsFunction(scope, 'FacilitiesDeleteByFcCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/facilities/_fcCollectionId/DELETE'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/facilities'),
+    handler: '_fcCollectionId/DELETE/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
