+++ lib/handlers/geozones/_gzCollectionId/GET/index.js
@@ -1,9 +1,9 @@
 const {
   getGeozonesByGzCollectionId,
   getGeozonesByGeozoneId,
-} = require("/opt/geozones/methods");
+} = require("../../methods");
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { ALLOWED_FILTERS } = require("/opt/geozones/configs");
+const { ALLOWED_FILTERS } = require("../../configs");
 
 /**
  * @api {get} /geozones/{gzCollectionId} GET
