+++ lib/handlers/geozones/_gzCollectionId/POST/index.js
@@ -1,7 +1,7 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
 const { quickApiPutHandler } = require("/opt/data-utils");
-const { GEOZONE_API_PUT_CONFIG } = require("/opt/geozones/configs");
-const { parseRequest } = require("/opt/geozones/methods");
+const { GEOZONE_API_PUT_CONFIG } = require("../../configs");
+const { parseRequest } = require("../../methods");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
 
 /**
