+++ lib/handlers/geozones/_gzCollectionId/PUT/index.js
@@ -1,7 +1,7 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
 const { quickApiUpdateHandler } = require("/opt/data-utils");
-const { GEOZONE_API_UPDATE_CONFIG } = require("/opt/geozones/configs");
-const { parseRequest } = require("/opt/geozones/methods");
+const { GEOZONE_API_UPDATE_CONFIG } = require("../../configs");
+const { parseRequest } = require("../../methods");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
 
 /**
