+++ lib/handlers/geozones/methods.js
@@ -5,7 +5,7 @@ const {
   incrementCounter
 } = require("/opt/dynamodb");
 const { Exception, logger } = require("/opt/base");
-const { ALLOWED_FILTERS } = require("/opt/geozones/configs");
+const { ALLOWED_FILTERS } = require("./configs");
 
 /**
  * Adds any filter expressions if any filters were added to query.
