+++ lib/handlers/geozones/resources.js
@@ -19,8 +19,8 @@ function geozonesSetup(scope, props) {
 
   // GET /geozones/{gzCollectionId}
   const geozonesGetByGzCollectionId = new NodejsFunction(scope, 'GeozonesGetByGzCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/geozones/_gzCollectionId/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/geozones'),
+    handler: '_gzCollectionId/GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -32,8 +32,8 @@ function geozonesSetup(scope, props) {
 
   // POST /geozones/{gzCollectionId}
   const geozonesPostByGzCollectionId = new NodejsFunction(scope, 'GeozonesPostByGzCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/geozones/_gzCollectionId/POST'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/geozones'),
+    handler: '_gzCollectionId/POST/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -42,8 +42,8 @@ function geozonesSetup(scope, props) {
 
   // PUT /geozones/{gzCollectionId}
   const geozonesPutByGzCollectionId = new NodejsFunction(scope, 'GeozonesPutByGzCollectionId', {
-    code: lambda.Code.fromAsset('lib/handlers/geozones/_gzCollectionId/PUT'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/geozones'),
+    handler: '_gzCollectionId/PUT/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -54,8 +54,8 @@ function geozonesSetup(scope, props) {
 
   // DELETE /geozones/{gzCollectionId}
   const geozonesDeleteByGeozoneId = new NodejsFunction(scope, 'GeozonesDeleteByGeozoneId', {
-    code: lambda.Code.fromAsset('lib/handlers/geozones/_gzCollectionId/DELETE'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/geozones'),
+    handler: '_gzCollectionId/DELETE/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
