+++ lib/handlers/products/_byPartition/GET/index.js
@@ -1,5 +1,5 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { getProducts } = require("/opt/products/methods");
+const { getProducts } = require("../../methods");
 
 exports.handler = async (event, context) => {
   logger.info("GET Products", event);
