+++ lib/handlers/products/methods.js
@@ -1,8 +1,8 @@
 
 const { TABLE_NAME, batchTransactData, runQuery, getOne, parallelizedBatchGetData } = require('/opt/dynamodb');
 const { Exception, buildDateRange, buildDateTimeFromShortDate, getNow, logger } = require('/opt/base');
-const { POLICY_TYPES } = require('/opt/policies/configs');
-const { PRODUCT_DAILY_PROPERTIES_CONFIG, PRODUCT_DEFAULT_PROPERTY_NAMES, PRODUCT_DEFAULT_SCHEDULE_RANGE } = require('/opt/products/configs');
+const { POLICY_TYPES } = require('../policies/configs');
+const { PRODUCT_DAILY_PROPERTIES_CONFIG, PRODUCT_DEFAULT_PROPERTY_NAMES, PRODUCT_DEFAULT_SCHEDULE_RANGE } = require('./configs');
 const { quickApiPutHandler } = require('/opt/data-utils');
 
 async function getProducts(acCollectionId, activityType, activityId, seasonId = null, productId = null, options = {}) {
