+++ lib/handlers/products/resources.js
@@ -18,8 +18,8 @@ function productsSetup(scope, props) {
 
   // GET /products/{acCollectionId}/{activityType}/{activityId} - Get all products for a given activity
   const productsGetMethod = new NodejsFunction(scope, 'ProductsGetByPartition', {
-    code: lambda.Code.fromAsset('lib/handlers/products/_byPartition/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers'),
+    handler: 'products/_byPartition/GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
