+++ lib/handlers/protectedAreas/POST/index.js
@@ -1,6 +1,6 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
 const { quickApiPutHandler } = require("/opt/data-utils");
-const { PROTECTED_AREA_API_PUT_CONFIG } = require("/opt/protectedAreas/configs");
+const { PROTECTED_AREA_API_PUT_CONFIG } = require("../configs");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
 
 exports.handler = async (event, context) => {
