+++ lib/handlers/protectedAreas/_orcs/GET/index.js
@@ -3,7 +3,7 @@
  * Fetch single protected area by ORCS
  */
 
-const { getProtectedAreaByOrcs } = require('/opt/protectedAreas/methods')
+const { getProtectedAreaByOrcs } = require('../../methods');
 const { logger, sendResponse } = require('/opt/base');
 
 exports.handler = async (event, context) => {
