+++ lib/handlers/protectedAreas/_orcs/PUT/index.js
@@ -1,6 +1,6 @@
 const { Exception, logger, sendResponse } = require("/opt/base");
 const { quickApiUpdateHandler } = require("/opt/data-utils");
-const { PROTECTED_AREA_API_UPDATE_CONFIG } = require("/opt/protectedAreas/configs");
+const { PROTECTED_AREA_API_UPDATE_CONFIG } = require("../../configs");
 const { TABLE_NAME, batchTransactData } = require("/opt/dynamodb");
 
 exports.handler = async (event, context) => {
