+++ lib/handlers/protectedAreas/configs.js
