+++ lib/handlers/protectedAreas/methods.js
