+++ lib/handlers/protectedAreas/resources.js
@@ -14,8 +14,8 @@ function protectedAreasSetup(scope, props) {
 
   // GET /protected-areas
   const protectedAreasGetFunction = new NodejsFunction(scope, 'ProtectedAreasGet', {
-    code: lambda.Code.fromAsset('lib/handlers/protectedAreas/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/protectedAreas'),
+    handler: 'GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -24,8 +24,8 @@ function protectedAreasSetup(scope, props) {
 
   // POST /protected-areas/
   const protectedAreasPostFunction = new NodejsFunction(scope, 'ProtectedAreasPost', {
-    code: lambda.Code.fromAsset('lib/handlers/protectedAreas/POST'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/protectedAreas'),
+    handler: 'POST/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -34,8 +34,8 @@ function protectedAreasSetup(scope, props) {
   
   // DELETE /protected-areas/
   const protectedAreasDeleteFunction = new NodejsFunction(scope, 'ProtectedAreasDelete', {
-    code: lambda.Code.fromAsset('lib/handlers/protectedAreas/DELETE'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/protectedAreas'),
+    handler: 'DELETE/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -47,8 +47,8 @@ function protectedAreasSetup(scope, props) {
 
   // GET /protected-areas/{orcs}
   const protectedAreasGetByOrcsFunction = new NodejsFunction(scope, 'ProtectedAreasGetByOrcs', {
-    code: lambda.Code.fromAsset('lib/handlers/protectedAreas/_orcs/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/protectedAreas'),
+    handler: '_orcs/GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -57,8 +57,8 @@ function protectedAreasSetup(scope, props) {
 
   // POST /protected-areas/{orcs}
   const protectedAreasPostByOrcsFunction = new NodejsFunction(scope, 'ProtectedAreasPostByOrcs', {
-    code: lambda.Code.fromAsset('lib/handlers/protectedAreas/_orcs/POST'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/protectedAreas'),
+    handler: '_orcs/POST/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -67,8 +67,8 @@ function protectedAreasSetup(scope, props) {
 
   // PUT /protected-areas/{orcs}
   const protectedAreasPutByOrcsFunction = new NodejsFunction(scope, 'ProtectedAreasPutByOrcs', {
-    code: lambda.Code.fromAsset('lib/handlers/protectedAreas/_orcs/PUT'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/protectedAreas'),
+    handler: '_orcs/PUT/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -77,8 +77,8 @@ function protectedAreasSetup(scope, props) {
   
   // DELETE /protected-areas/{orcs}
   const protectedAreasDeleteByOrcsFunction = new NodejsFunction(scope, 'ProtectedAreasDeleteByOrcs', {
-    code: lambda.Code.fromAsset('lib/handlers/protectedAreas/_orcs/DELETE'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/protectedAreas'),
+    handler: '_orcs/DELETE/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
