+++ lib/handlers/transactions/GET/index.js
@@ -1,7 +1,7 @@
 // Get transaction
 
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { getTransactionByTransactionId } = require("/opt/transactions/methods");
+const { getTransactionByTransactionId } = require("../methods");
 
 exports.handler = async (event, context) => {
   logger.info("Transactions GET:", event);
