+++ lib/handlers/transactions/POST/index.js
@@ -1,9 +1,9 @@
 // Create new transaction
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { createTransaction } = require("/opt/transactions/methods");
+const { createTransaction } = require("../methods");
 const { batchTransactData, TABLE_NAME } = require("/opt/dynamodb");
 const { quickApiPutHandler } = require("/opt/data-utils");
-const { TRANSACTION_PUT_CONFIG } = require("/opt/transactions/configs");
+const { TRANSACTION_PUT_CONFIG } = require("../configs");
 
 exports.handler = async (event, context) => {
   logger.info("Transactions POST:", event);
