+++ lib/handlers/transactions/resources.js
@@ -14,8 +14,8 @@ function transactionsSetup(scope, props) {
 
   // GET /transactions
   const transactionsGet = new NodejsFunction(scope, 'TransactionsGet', {
-    code: lambda.Code.fromAsset('lib/handlers/transactions/GET'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/transactions'),
+    handler: 'GET/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
@@ -25,8 +25,8 @@ function transactionsSetup(scope, props) {
 
   // POST /transactions
   const transactionsPost = new NodejsFunction(scope, 'TransactionsPost', {
-    code: lambda.Code.fromAsset('lib/handlers/transactions/POST'),
-    handler: 'index.handler',
+    code: lambda.Code.fromAsset('lib/handlers/transactions'),
+    handler: 'POST/index.handler',
     runtime: lambda.Runtime.NODEJS_20_X,
     environment: props.env,
     layers: props.layers,
