+++ lib/handlers/worldlineNotification/POST/index.js
@@ -1,16 +1,16 @@
 // Await Worldline Notification push
 const querystring = require('querystring');
 const { Exception, logger, sendResponse } = require("/opt/base");
-const { updateTransaction } = require("/opt/transactions/methods");
+const { updateTransaction } = require("../../transactions/methods");
 const { batchTransactData, TABLE_NAME } = require("/opt/dynamodb");
 const { quickApiUpdateHandler } = require("/opt/data-utils");
-const { TRANSACTION_UPDATE_CONFIG } = require("/opt/transactions/configs");
+const { TRANSACTION_UPDATE_CONFIG } = require("../../transactions/configs");
 
 const WORLDLINE_WEBHOOK_SECRET = process.env.WORLDLINE_WEBHOOK_SECRET || 'default-secret'
 
 // Update booking
-const { confirmBooking } = require("/opt/bookings/methods")
-const { BOOKING_UPDATE_CONFIG } = require("/opt/bookings/configs");
+const { confirmBooking } = require("../../bookings/methods")
+const { BOOKING_UPDATE_CONFIG } = require("../../bookings/configs");
 
 exports.handler = async (event, context) => {
   logger.info("Worldline Notification POST:", event);
