+++ lib/handlers/worldlineNotification/resources.js
@@ -13,8 +13,8 @@ function worldlineNotificationSetup(scope, props) {
 
   // POST /worldline-notification
   const worldlineNotificationPost = new NodejsFunction(scope, "WorldlineNotificationPost", {
-      code: lambda.Code.fromAsset("lib/handlers/worldlineNotification/POST"),
-      handler: "index.handler",
+      code: lambda.Code.fromAsset("lib/handlers"),
+      handler: "worldlineNotification/POST/index.handler",
       runtime: lambda.Runtime.NODEJS_20_X,
       environment: props.env,
       layers: props.layers,
